// src/sdk/types.gen.ts
export type Address = {
    readonly uuid: string;
    line1?: string | null;
    line2?: string | null;
    city?: string | null;
    state?: string | null;
    zipcode?: string | null;
    country?: string | null;
};
// ...
export type BlankEnum = '';
// ...
export type GroupHome = {
    readonly uuid: string;
    name: string;
    timezone?: TimezoneEnum | BlankEnum | NullEnum | null;
    phone: string;
    fax: string;
    email?: string | null;
    emergency_contact_number?: string | null;
    address?: Address;
    license?: License;
    no_of_rooms: number;
    active?: boolean;
    shifts?: Array<GroupHomeShift>;
    readonly rooms: Array<GroupHomeRoom>;
    readonly media: string;
    readonly avatar_url: string;
    readonly created_at: string | null;
    readonly updated_at: string | null;
};

export type GroupHomeRoom = {
    readonly uuid: string;
    room_number: string;
    is_active?: boolean;
    is_occupied?: boolean;
};

export type GroupHomeShift = {
    readonly uuid: string;
    shift: ShiftEnum;
    start_time: string;
    end_time: string;
    is_active?: boolean;
};
// ...
export type License = {
    readonly uuid: string;
    number: string;
    start_date?: string | null;
    expiry_date?: string | null;
    readonly created_at: string | null;
    readonly updated_at: string | null;
};
// ...
export type NullEnum = unknown;
// ...
export type ShiftEnum = 'MORNING' | 'EVENING' | 'NIGHT';
// ...
export type TimezoneEnum = 'Eastern Time' | 'Central Time' | 'Mountain Time' | 'Pacific Time' | 'Arizona Time' | 'Alaska Time';
// ...
export type AddressWritable = {
    line1?: string | null;
    line2?: string | null;
    city?: string | null;
    state?: string | null;
    zipcode?: string | null;
    country?: string | null;
};
// ...
export type GroupHomeWritable = {
    name: string;
    timezone?: TimezoneEnum | BlankEnum | NullEnum | null;
    phone: string;
    fax: string;
    email?: string | null;
    emergency_contact_number?: string | null;
    address?: AddressWritable;
    license?: LicenseWritable;
    no_of_rooms: number;
    active?: boolean;
    shifts?: Array<GroupHomeShiftWritable>;
    certificate_media_ids?: Array<string>;
    image_media_id?: string | null;
};

export type GroupHomeRoomWritable = {
    room_number: string;
    is_active?: boolean;
    is_occupied?: boolean;
};

export type GroupHomeShiftWritable = {
    shift: ShiftEnum;
    start_time: string;
    end_time: string;
    is_active?: boolean;
};
// ...
export type LicenseWritable = {
    number: string;
    start_date?: string | null;
    expiry_date?: string | null;
};